Time: 16 ms (59.81%) | Memory: 18.3 MB (42.71%) - LeetSync

+++ 2163-kth-distinct-string-in-an-array/kth-distinct-string-in-an-array.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     string kthDistinct(vector<string>& arr, int k) {
     map<string, int> b;
-    int n = arr.size();
+    long n = arr.size();
     for (int i = 0; i < n; i++) {
         b[arr[i]]++;
     }
